Create Transmit_Threshold.py
File is not complete. Implemented a threshold for a frequency of 5 kHz using an fft and the packages scipy and numpy. Need to pull the voltage data (hydrophone readings) from the sampled text file and perform the fft on the voltage list. Idea is to only initiate transmit() function if the threshold is met. If the threshold is not met, the data will not be sent to the server however it will still be saved to the usb flash drive aboard the buoy on the raspberry pi.